fix(account): drop the duplicated purchase history from the account page
The account page carried a purchase-history tab alongside the top-up one, but
the same table already lives on the shop page, where buying and reading the
receipt happen in the same place. Two copies of one table only made the page
that regular users land on longer.

With one record type left there is nothing to tab between, so the card renders
the table directly. The component is renamed AccountRecords -> RedeemHistory:
the old name promised more than it now shows, and a reviewer reading it would
expect several kinds of record.

/user/orders is untouched — the shop still reads it. Only the account page's
second request and its table are gone.
